% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneCentricNonCoding_preload.R
\name{GeneCentricNonCoding_preload}
\alias{GeneCentricNonCoding_preload}
\title{Title Gene-centric analysis of noncoding functional categories using SurvSTAAR procedure}
\usage{
GeneCentricNonCoding_preload(
  gene_name,
  genofile,
  objNull,
  variant_type = c("SNV", "Indel", "variant"),
  genes_info,
  categories = c("all", "downstream", "upstream", "UTR", "promoter_CAGE", "promoter_DHS",
    "enhancer_CAGE", "enhancer_DHS"),
  dfPromCAGEVarGene.SNV,
  variant.id.SNV.PromCAGE,
  dfPromrOCRsVarGene.SNV,
  variant.id.SNV.PromrOCRs,
  dfHancerCAGEVarGene.SNV,
  variant.id.SNV.HancerCAGE,
  dfHancerrOCRsVarGene.SNV,
  variant.id.SNV.HancerrOCRs,
  rare_maf_cutoff = 0.01,
  rare_num_cutoff = 2,
  geno_missing_cutoff = 1,
  geno_missing_imputation = c("mean", "minor"),
  min_maf_cutoff = 0,
  combine_ultra_rare = TRUE,
  ultra_rare_mac_cutoff = 20,
  QC_label = "annotation/filter",
  Annotation_dir = "annotation/info/FunctionalAnnotation",
  Annotation_name_catalog,
  Use_annotation_weights = TRUE,
  Annotation_name = NULL,
  use_SPA = NULL,
  SPA_filter = TRUE,
  SPA_filter_cutoff = 0.05,
  rm_long = TRUE,
  rm_long_cutoff = 5000,
  verbose = FALSE
)
}
\arguments{
\item{gene_name}{a character string specifying the name of the gene to be analyzed.}

\item{genofile}{an object of opened annotated GDS (aGDS) file.}

\item{objNull}{an object from null model, which is the output of \code{nullModel}.}

\item{variant_type}{the type of variant included in the analysis.
Options are \code{“SNV”}, \code{“Indel”}, or \code{“variant”} (default = “SNV”).}

\item{genes_info}{a matrix containing information on gene names and chromosomes.}

\item{categories}{the noncoding functional category to be analyzed using SurvSTAAR
procedure. Choices include \code{all}, \code{downstream}, \code{upstream}, \code{UTR},
\code{promoter_CAGE}, \code{promoter_DHS}, \code{enhancer_CAGE}, \code{enhancer_DHS}
(default = "all").}

\item{dfPromCAGEVarGene.SNV}{a data frame containing preloaded information for
\code{promoter_CAGE}.}

\item{variant.id.SNV.PromCAGE}{a vector containing variant IDs for
\code{promoter_CAGE}.}

\item{dfPromrOCRsVarGene.SNV}{a data frame containing preloaded information for
\code{promoter_DHS}.}

\item{variant.id.SNV.PromrOCRs}{a vector containing variant IDs for
\code{promoter_DHS}.}

\item{dfHancerCAGEVarGene.SNV}{a data frame containing preloaded information for
\code{enhancer_CAGE}.}

\item{variant.id.SNV.HancerCAGE}{a vector containing variant IDs for
\code{enhancer_CAGE}.}

\item{dfHancerrOCRsVarGene.SNV}{a data frame containing preloaded information for
\code{enhancer_DHS}.}

\item{variant.id.SNV.HancerrOCRs}{a vector containing variant IDs for
\code{enhancer_DHS}.}

\item{rare_maf_cutoff}{a numeric value indicating the maximum minor allele frequency
for defining rare variants (default = 0.01).}

\item{rare_num_cutoff}{a numeric value indicating the minimum number of variants
required for analyzing a given variant set (default = 2).}

\item{geno_missing_cutoff}{a numeric value specifying the cutoff for missing rates.
Variants with a missing rate higher than this cutoff are excluded (default = 1,
do not remove any variants).}

\item{geno_missing_imputation}{a character string specifying the method for
handling missing genotypes. Options are “mean” or “minor” (default = “mean”).}

\item{min_maf_cutoff}{a numeric variable representing the minimum minor allele
frequency (MAF) to include in the individual test (default = 0).}

\item{combine_ultra_rare}{a logical variable indicating whether to combine ultra-rare
variants in the SKAT and ACAT-V test under imbalanced phenotypes (default = TRUE).}

\item{ultra_rare_mac_cutoff}{a numeric value indicating the maximum MAC for
ultra-rare variants (default = 20).}

\item{QC_label}{the channel name of the QC label in the GDS/aGDS file
(default = "annotation/filter").}

\item{Annotation_dir}{the channel name of the annotations in the aGDS file
(default = "annotation/info/FunctionalAnnotation").}

\item{Annotation_name_catalog}{a data frame containing annotation names and their
corresponding channel names in the aGDS file.}

\item{Use_annotation_weights}{a logical variable indicating whether to use annotations
as weights (default = TRUE).}

\item{Annotation_name}{a vector of annotation names used in SurvSTAAR (default = NULL).}

\item{use_SPA}{a logical value indicating whether to use SaddlePoint
Approximation (SPA) in the analysis. Highly recommended for imbalanced
phenotypes (default = TRUE).}

\item{SPA_filter}{a logical value indicating whether to use SPA to calibrate
P-values for imbalanced traits when the P-value is below a pre-specified
threshold (default = TRUE).}

\item{SPA_filter_cutoff}{a numeric value specifying the threshold for recalculating
P-values using the SPA method (default = 0.05).}

\item{rm_long}{a logical value indicating whether to remove the long masks
during the analysis (default = TRUE).}

\item{rm_long_cutoff}{a numeric value specifying the cutoff for the number of
variants in a mask to remove (default = 5000).}

\item{verbose}{a logical value indicating whether to provides additional detailed
information or messages during the execution of this function (default = FALSE).}
}
\value{
a list of results containing the SurvSTAAR results.
\code{gene_info} is the information about the analyzed gene name and chromosome.
\code{SurvSTAAR_O_all} is the p-values of the functional categories from the SurvSTAAR-O test.
Other data frames are the detailed results for each functional category.
}
\description{
The \code{GeneCentricNonCoding_preload} function takes in chromosome, gene name,
functional category, the object of opened annotated GDS file, and the object from null
model to analyze the association between a time-to-event phenotype and noncoding
functional categories of a gene by using SurvSTAAR procedure.
}
\details{
For each noncoding functional category, the SurvSTAAR-O p-value is derived from an
omnibus test that combines SKAT(1,25), SKAT(1,1), Burden(1,25), Burden(1,1),
ACAT-V(1,25), and ACAT-V(1,1), with p-values weighted by each annotation using
the Cauchy combination method.
}
