% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndividualTestGDS.R
\name{IndividualTestGDS}
\alias{IndividualTestGDS}
\title{Score test for individual variants in a given region of an annotated GDS (aGDS) format file}
\usage{
IndividualTestGDS(
  objNull,
  start_loc,
  end_loc,
  genofile,
  chr,
  rs_channel = NULL,
  QC_label = "annotation/filter",
  variant_type = c("variant", "SNV", "Indel"),
  use_SPA = NULL,
  SPA_filter = TRUE,
  SPA_filter_cutoff = 0.05,
  geno_missing_cutoff = 1,
  geno_missing_imputation = c("mean", "minor"),
  min_mac_cutoff = NULL,
  min_maf_cutoff = 0.01,
  chunk_size = 5000,
  verbose = FALSE
)
}
\arguments{
\item{objNull}{an object from null model, which is the output of \code{nullModel}.}

\item{start_loc}{starting location (position) of the genetic region for each
individual variant to be analyzed using score test.}

\item{end_loc}{ending location (position) of the genetic region for each
individual variant to be analyzed using score test.}

\item{genofile}{an SeqVarGDSClass object of opened annotated GDS (aGDS) file.}

\item{chr}{a numeric value representing the chromosome.}

\item{rs_channel}{channel name of the rsID in the aGDS file. Can be NULL.}

\item{QC_label}{channel name of the QC label in the aGDS file (default = "annotation/filter").}

\item{variant_type}{type of variant included in the analysis. Choices include
"variant", "SNV", or "Indel" (default = "variant").}

\item{use_SPA}{a logical value indicating whether to use SaddlePoint
Approximation (SPA) in the analysis. Highly recommended for imbalanced
phenotypes (default = TRUE).}

\item{SPA_filter}{a logical value indicating whether to use SPA to calibrate
P-values for imbalanced traits when the P-value is below a pre-specified
threshold (default = TRUE).}

\item{SPA_filter_cutoff}{A numeric value specifying the threshold for recalculating
P-values using the SPA method (default = 0.05).}

\item{geno_missing_cutoff}{a numeric value specifying the cutoff for missing rates.
Variants with a missing rate higher than this cutoff are excluded (default = 1,
do not remove any variants).}

\item{geno_missing_imputation}{a character string specifying the method for
handling missing genotypes. Options are “mean” or “minor” (default = “mean”).}

\item{min_mac_cutoff}{a numeric variable representing the minimum minor allele
count (MAC) to include in the individual test. Can be NULL.}

\item{min_maf_cutoff}{a numeric variable representing the minimum minor allele
frequency (MAF) to include in the individual test (default = 0.01).}

\item{chunk_size}{a numeric variable representing the number of variants analyzed
in a single chunk (default = 1000).}

\item{verbose}{a logical value indicating whether to provides additional detailed
information or messages during the execution of this function (default = FALSE).}
}
\value{
A data frame containing the score test results and the estimated effect
size of the minor allele for each individual variant in the specified genotype file.
The columns include: CHR (chromosome), MarkerID (rsID), POS (position),
REF (reference allele), ALT (alternative allele), Pvalue (P-value estimated
from Chi-squared distribution),	Pvalue_SPA (P-value approximated using the SPA),
pvalue_log10 (-log10(P-value). If \code{use_SPA = TRUE}, this is calculated
using Pvalue_SPA; otherwise, it is based on Pvalue.), Score (estimated score
statistic), Variance (variance of the score statistic), Stest (Score^2 / Variance),
Est (estimated effect), Est_se (standard error of the estimated effect size),
MAF (minor allele frequency), MAC (minor allele count), missing_rate (missing rate).
}
\description{
The \code{IndividualTestGDS} performs association analysis between a
time-to-event (TTE) phenotype and individual variants in a aGDS file using
the exact score test.
}
