% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndividualTestPlink.R
\name{IndividualTestPlink}
\alias{IndividualTestPlink}
\title{Score test for individual variants in a given plink format file}
\usage{
IndividualTestPlink(
  objNull,
  genofile,
  chr,
  sampleCol,
  use_SPA = NULL,
  SPA_filter = TRUE,
  SPA_filter_cutoff = 0.05,
  geno_missing_cutoff = 1,
  geno_missing_imputation = c("mean", "minor"),
  min_mac_cutoff = NULL,
  min_maf_cutoff = 0.01,
  chunk_size = 1000,
  verbose = FALSE
)
}
\arguments{
\item{objNull}{an object from null model, which is the output of \code{nullModel}.}

\item{genofile}{a character string specifying the file name of the genotype file.}

\item{chr}{a numeric value representing the chromosome.}

\item{sampleCol}{a character string specifying the column name of sample IDs.}

\item{use_SPA}{a logical value indicating whether to use SaddlePoint
Approximation (SPA) in the analysis. Highly recommended for imbalanced
phenotypes (default = TRUE).}

\item{SPA_filter}{a logical value indicating whether to use SPA to calibrate
P-values for imbalanced traits when the P-value is below a pre-specified
threshold (default = TRUE).}

\item{SPA_filter_cutoff}{A numeric value specifying the threshold for recalculating
P-values using the SPA method (default = 0.05).}

\item{geno_missing_cutoff}{a numeric value specifying the cutoff for missing rates.
Variants with a missing rate higher than this cutoff are excluded (default = 1,
do not remove any variants).}

\item{geno_missing_imputation}{a character string specifying the method for
handling missing genotypes. Options are “mean” or “minor” (default = “mean”).}

\item{min_mac_cutoff}{a numeric variable representing the minimum minor allele
count (MAC) to include in the individual test. Can be NULL.}

\item{min_maf_cutoff}{a numeric variable representing the minimum minor allele
frequency (MAF) to include in the individual test (default = 0.01).}

\item{chunk_size}{a numeric variable representing the number of variants analyzed
in a single chunk (default = 1000).}

\item{verbose}{a logical value indicating whether to provides additional detailed
information or messages during the execution of this function (default = FALSE).}
}
\value{
A data frame containing the score test results and the estimated effect
size of the minor allele for each individual variant in the specified genotype file.
The columns include: CHR (chromosome), MarkerID (rsID), POS (position),
REF1 (alternative allele), REF2 (reference allele), Pvalue (P-value estimated
from Chi-squared distribution),	Pvalue_SPA (P-value approximated using the SPA),
pvalue_log10 (-log10(P-value). If \code{use_SPA = TRUE}, this is calculated
using Pvalue_SPA; otherwise, it is based on Pvalue.), Score (estimated score
statistic), Variance (variance of the score statistic), Stest (Score^2 / Variance),
Est (estimated effect), Est_se (standard error of the estimated effect size),
MAF (minor allele frequency), MAC (minor allele count), missing_rate (missing rate).
}
\description{
The \code{IndividualTestPlink} performs association analysis between a
time-to-event (TTE) phenotype and individual variants in a PLINK file using
the exact score test.
}
