% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NullModel.R
\name{NullModel}
\alias{NullModel}
\title{Fit Cox proportional hazards model under the null hypothesis for related samples.}
\usage{
NullModel(
  genofile = NULL,
  phenofile,
  LOCO = TRUE,
  chr = NULL,
  statusCol,
  timeCol,
  sampleCol,
  covCol = NULL,
  PRSCol = NULL,
  use_SPA = FALSE,
  range = c(-100, 100),
  length.out = 10000,
  verbose = FALSE
)
}
\arguments{
\item{genofile}{a genome file, which can be NULL. If not NULL, it can either
be a character string specifying the file name in PLINK format or a SeqVarGDSClass
object for aGDS format data. When this parameter is provided (non-NULL), the
function will match sample IDs with genome IDs; if it is NULL, no matching
will be performed.}

\item{phenofile}{a character string specifying the file name of the phenotype
data, which includes all variables used in the null model.}

\item{LOCO}{a logical value indicating whether to perform leave-one-chromosome-out (LOCO)
in null model fitting (default = TRUE).}

\item{chr}{a numeric value representing the chromosome. If `LOCO = FALSE`, this
can be NULL. If `LOCO = TRUE`, `chr` must be specified as a valid chromosome number.}

\item{statusCol}{the column name of survival status.}

\item{timeCol}{the column name of survival time.}

\item{sampleCol}{the column name of sample IDs.}

\item{covCol}{a vector of column names of covariates.}

\item{PRSCol}{the column name of polygenic risk scores (PRS). If `LOCO = FALSE`,
it is recommended to use "chr23".}

\item{use_SPA}{a logical value indicating whether to use SPA (SaddlePoint
Approximation) in the subsequent analysis, it is highly recommended under
imbalanced phenotypes (default = TRUE).}

\item{range}{a two-element numeric vector to specify the domain of the
empirical CGF (default = c(-100,100)).}

\item{length.out}{a positive integer for empirical CGF. Larger length.out
corresponds to longer calculation time and more accurate estimated empirical
CGF (default = 1e4).}

\item{verbose}{a logical value indicating whether to provides additional detailed
information or messages during the execution of this function (default = FALSE).}
}
\value{
A list containing the model fit from \code{\link{coxph}}.
If \code{use_SPA = TRUE}, the list will also include the empirical CGF.
}
\description{
This function fit Cox proportional hazards null model using the \code{\link{coxph}}
function from the \code{\link{survival}} package. Tied event times are handled
using Breslow’s approximation. Additionally, the function calculates the empirical
cumulative generating function (CGF) of the martingale residuals using the
\code{CGF4MartingaleRes} function from the \code{SPACox} package.
}
