% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvSTAAR-package.R
\docType{package}
\name{SurvSTAAR-package}
\alias{SurvSTAAR-package}
\title{SurvSTAAR: SurvSTAAR for Analyzing Whole-Genome/Whole-Exome Sequencing Data in Time-to-Event Studies}
\description{
An R package for performing SurvSTAAR in analyzing whole-genome/whole-exome sequencing data in time-to-event studies.
}
\keyword{internal}
