% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvSTAAR.R
\name{SurvSTAAR}
\alias{SurvSTAAR}
\title{STAAR procedure using omnibus test}
\usage{
SurvSTAAR(
  Geno,
  MAF = NULL,
  MAC = NULL,
  objNull,
  annotation_phred = NULL,
  rare_maf_cutoff = 0.01,
  rare_num_cutoff = 2,
  combine_ultra_rare = TRUE,
  ultra_rare_mac_cutoff = 20,
  use_SPA = NULL,
  SPA_filter = TRUE,
  SPA_filter_cutoff = 0.05,
  verbose = FALSE
)
}
\arguments{
\item{Geno}{an n*p genotype matrix (dosage matrix) of the target sequence,
where n is the sample size and p is the number of genetic variants.}

\item{MAF}{minor allele frequency, can be NULL.}

\item{MAC}{minor allele count, can be NULL.}

\item{objNull}{an object from null model, which is the output of \code{nullModel}.}

\item{annotation_phred}{a data frame or matrix of functional annotation data
of dimension p*q (or a vector of a single annotation score with length p).}

\item{rare_maf_cutoff}{a numeric value indicating the maximum minor allele frequency
for defining rare variants (default = 0.01).}

\item{rare_num_cutoff}{a numeric value indicating the minimum number of variants
required for analyzing a given variant set (default = 2).}

\item{combine_ultra_rare}{a logical variable indicating whether to combine ultra-rare
variants in the SKAT and ACAT-V test under imbalanced phenotypes (default = TRUE).}

\item{ultra_rare_mac_cutoff}{a numeric value indicating the maximum MAC for
ultra-rare variants (default = 20).}

\item{use_SPA}{a logical value indicating whether to use SaddlePoint
Approximation (SPA) in the analysis. Highly recommended for imbalanced
phenotypes (default = TRUE).}

\item{SPA_filter}{a logical value indicating whether to use SPA to calibrate
P-values for imbalanced traits when the P-value is below a pre-specified
threshold (default = TRUE).}

\item{SPA_filter_cutoff}{a numeric value specifying the threshold for recalculating
P-values using the SPA method (default = 0.05).}

\item{verbose}{a logical value indicating whether to provides additional detailed
information or messages during the execution of this function (default = FALSE).}
}
\value{
A list containing the set-based test results from \code{SurvSTAAR.}
\code{num_variant} is the number of rare variants in this analyzed set.
\code{MAF} is minor allele frequency.
\code{MAC} is minor allele count.
}
\description{
The \code{SurvSTAAR} function takes in genotype, the object from fitting the
null model, and functional annotation data to analyze the association between
a time-to-event phenotype and a variant-set by using SurvSTAAR procedure.
}
